rag_evaluator/requirements.txt
ragas==0.2.7
langsmith>=0.0.77,<0.2.0
langchain-core>=0.1.0
langchain-community>=0.0.38
langchain>=0.0.350,<0.2.0
openai>=1.3.0,<2.0.0
datasets>=2.14.0,<3.0.0
python-dotenv>=1.0.0
pandas==2.1.4
matplotlib==3.8.2
plotly==5.18.0
